+++ src/Resource/InvoiceResource.php
@@ -7,6 +7,7 @@ namespace Freelo\Sdk\Resource;
 use Freelo\Sdk\Exception\ApiException;
 use Freelo\Sdk\Http\PaginatedResult;
 use Freelo\Sdk\Model\IssuedInvoice;
+use Freelo\Sdk\Model\WorkReportExtended;
 
 /**
  * Invoice resource manager
@@ -50,6 +51,20 @@ class InvoiceResource extends AbstractResource
         $data = $this->parser->parseSingle($response);
 
         return IssuedInvoice::fromArray($data);
+    }
+
+    /**
+     * Get invoice reports as JSON
+     *
+     * @return WorkReportExtended[]
+     * @throws ApiException
+     */
+    public function getReportsJson(int $invoiceId): array
+    {
+        $response = $this->client->get("issued-invoice/{$invoiceId}/reports-json");
+        $data = $this->parser->parseCollection($response);
+
+        return array_map(fn(array $item) => WorkReportExtended::fromArray($item), $data);
     }
 
     /**